requirements.txt
Django==4.2.7
asgiref==3.7.2
sqlparse==0.4.4
tzdata==2023.3

# REST API
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.57.0

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# Performance & Caching
django-redis==5.4.0
redis==5.0.1

# File Handling & Export
Pillow==10.1.0
openpyxl==3.1.2
reportlab==4.0.7

# API Documentation
drf-spectacular==0.27.0

# Monitoring & Logging
django-debug-toolbar==4.2.0

# Task Queue
celery==5.3.4

# Utilities
django-extensions==3.2.3
django-crispy-forms==2.1
crispy-bootstrap5==2024.2

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0
